Add configuration and code for xDS flow control validation demo
