+++ xds-demo/product/script/mysql.yaml
