+++ xds-demo/product/script/spring-client.yaml
