+++ xds-demo/product/script/spring-server-destination.yaml
