+++ xds-demo/product/script/spring-server-virtureservice-flowcontrol.yaml
@@ -1,0 +1,27 @@
+apiVersion: networking.istio.io/v1alpha3
+kind: VirtualService
+metadata:
+  name: spring-server-virtualservice
+spec:
+  hosts:
+  - spring-server
+  http:
+  - name: "v1-routes"
+    match:
+    - headers:
+        version:
+          exact: v1
+      uri:
+        exact: /router
+      ignoreUriCase: false
+    route:
+    - destination:
+        host: spring-server
+        subset: v1
+        port:
+          number: 8081
+    fault:
+      abort:
+        percentage:
+          value: 100
+        httpStatus: 400
