+++ xds-demo/product/script/spring-server-virtureservice.yaml
