+++ xds-demo/product/script/spring-server.yaml
