+++ xds-demo/product/spring-client/Dockerfile
