+++ xds-demo/product/spring-client/build-client.sh
