+++ xds-demo/product/spring-server/Dockerfile
