+++ xds-demo/product/spring-server/build-server.sh
