+++ xds-demo/product/spring-server/start.sh
