--- xds-service-discovery-demo/config/bootstrap.properties
@@ -1,8 +1,0 @@
-# Specifies the artifact name of the current Sermant artifact, used to identify different Sermant artifacts in a multi-Sermant scenario, and to isolate different Sermant configurations and service governance capabilities based on the artifact
-artifact=default
-# Specifies application name of the current host instance accessed by Sermant
-appName=default
-# Specifies application type of the current host instance accessed by Sermant
-appType=default
-# Specifies service name of the current host instance accessed by Sermant
-serviceName=default
