--- xds-service-discovery-demo/config/logback.xml
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<configuration scan="true" debug="false">
-    <!--  No operation event listener  -->
-    <statusListener class="ch.qos.logback.core.status.NopStatusListener"/>
-
-    <!-- Configure the log file output location -->
-    <property name="log.home_dir" value="${sermant_log_dir:-./logs/sermant/core}"/>
-    <property name="log.app_name" value="${sermant_app_name:-sermant}"/>
-    <!-- The maximum log history -->
-    <property name="log.maxHistory" value="${sermant_log_max_history:-30}"/>
-    <property name="log.level" value="${sermant_log_level:-info}"/>
-    <property name="log.maxSize" value="${sermant_log_max_size:-50MB}"/>
-    <property name="log.totalSize" value="${sermant_log_total_size:-20GB}"/>
-
-    <!-- Configure the log output format -->
-    <!-- %d{yyyy-MM-dd HH:mm:ss.SSS} for date-->
-    <!-- %C for class name-->
-    <!-- %F for file name-->
-    <!-- %M for method name-->
-    <!-- %L for line number-->
-    <!-- %thread for thread name-->
-    <!-- %m %msg for message-->
-    <!-- %n for line break-->
-    <property name="log.pattern" value="%d{yyyy-MM-dd HH:mm:ss.SSS} [%le] [%C] [%M:%L] [%thread] %m%n"/>
-
-    <!-- ConsoleAppender for console output log -->
-    <appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
-        <filter class="ch.qos.logback.classic.filter.LevelFilter">
-            <level>ERROR</level>
-            <onMatch>ACCEPT</onMatch>
-            <onMismatch>DENY</onMismatch>
-        </filter>
-        <encoder>
-            <pattern>
-                ${log.pattern}
-            </pattern>
-        </encoder>
-    </appender>
-
-    <!--Configure an up-pass appender that outputs logs of all levels-->
-    <appender name="app" class="ch.qos.logback.core.rolling.RollingFileAppender">
-        <rollingPolicy class="ch.qos.logback.core.rolling.SizeAndTimeBasedRollingPolicy">
-            <fileNamePattern>${log.home_dir}/app/%d{yyyy-MM-dd}/${log.app_name}-%i.log</fileNamePattern>
-            <maxHistory>${log.maxHistory}</maxHistory>
-            <MaxFileSize>${log.maxSize}</MaxFileSize>
-            <totalSizeCap>${log.totalSize}</totalSizeCap>
-            <cleanHistoryOnStart>true</cleanHistoryOnStart>
-        </rollingPolicy>
-        <encoder>
-            <pattern>
-                ${log.pattern}
-            </pattern>
-        </encoder>
-    </appender>
-
-    <root>
-        <!-- Print logs for debug and higher level -->
-        <level value="${log.level}"/>
-        <appender-ref ref="app"/>
-    </root>
-
-</configuration>
