--- xds-service-discovery-demo/config/plugins.yaml
@@ -1,17 +1,0 @@
-# The premain loads the plugin directory configured under plugins, and the agentmain loads the active plugin directory under dynamicPlugins
-# Plugins are loaded in the following order:
-# 1. First load the default plugins in the order configured in 'plugins'.
-# 2. Then load the list of scene plugins according to the scene order configured in 'profile'. The loading order of scene plugins is the same as that in the configuration file.
-# 3. If the plugin configured in 'profiles' has been loaded before, it will not be loaded again.
-# 4. Bytecode enhancements at the same intercept point for different plugins take effect in the same order as the plugins load.
-# 'plugins' is used to configure static plugins. The plugins take effect when Sermant starts by premain. Plugins that support static installation are configured and cannot be uninstalled
-plugins:
-  - xds-service-discovery
-# 'dynamicPlugins' is used to configure plugins that support dynamic installation. The plugins take effect when Sermant starts by agentmain. allows uninstallation:
-# 1. active plugins will be actively enabled
-# 2. passive plugins need to be enabled by commands
-dynamicPlugins:
-  active:
-#    - active-plugin
-  passive:
-#    - passive-plugin
