--- xds-service-discovery-demo/pom.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <groupId>io.sermant</groupId>
-    <artifactId>xds-service-discovery-demo</artifactId>
-    <version>2.0.0</version>
-
-    <modules>
-        <module>spring-client</module>
-        <module>spring-server</module>
-        <module>xds-service-discovery</module>
-    </modules>
-
-    <packaging>pom</packaging>
-
-    <properties>
-        <maven.compiler.source>8</maven.compiler.source>
-        <maven.compiler.target>8</maven.compiler.target>
-        <spring.version>2.7.15</spring.version>
-    </properties>
-
-</project>
