--- xds-service-discovery-demo/product/spring-client/build-client.sh
@@ -1,1 +1,0 @@
-docker build -f Dockerfile -t spring-client:1.0.0 .
