--- xds-service-discovery-demo/product/spring-client/spring-client.yaml
@@ -1,35 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: spring-client
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: spring-client
-  template:
-    metadata:
-      labels:
-        app: spring-client
-    spec:
-      containers:
-      - name: spring-client
-        image: spring-client:1.0.0
-        imagePullPolicy: IfNotPresent
-        ports:
-        - containerPort: 8080
-      imagePullSecrets:
-        - name: default-secret
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: spring-client
-spec:
-  type: NodePort
-  ports:
-  - port: 8080
-    targetPort: 8080
-    nodePort: 30110
-  selector:
-    app: spring-client
