--- xds-service-discovery-demo/product/spring-client/start.sh
@@ -1,1 +1,0 @@
-exec java -javaagent:/home/agent/sermant-agent.jar -jar /home/spring-client.jar
