--- xds-service-discovery-demo/product/spring-server/Dockerfile
@@ -1,6 +1,0 @@
-FROM openjdk:8
-WORKDIR /home
-COPY spring-server.jar /home/spring-server.jar
-COPY start.sh /home
-RUN chmod -R 777 /home
-ENTRYPOINT ["sh", "/home/start.sh"]
