--- xds-service-discovery-demo/product/spring-server/build-server.sh
@@ -1,1 +1,0 @@
-docker build -f Dockerfile -t spring-server:1.0.0 .
