--- xds-service-discovery-demo/product/spring-server/start.sh
@@ -1,1 +1,0 @@
-exec java -jar /home/spring-server.jar
