--- xds-service-discovery-demo/spring-client/src/main/java/io/demo/spring/client/SpringClientApplication.java
@@ -1,33 +1,0 @@
-/*
- * Copyright (C) 2024-2024 Sermant Authors. All rights reserved.
- *
- *   Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- */
-
-package io.demo.spring.client;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * SpringClientApplication
- *
- * @author daizhenyu
- * @since 2024-07-02
- **/
-@SpringBootApplication
-public class SpringClientApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(SpringClientApplication.class, args);
-    }
-}
