--- xds-service-discovery-demo/spring-server/src/main/java/io/demo/spring/server/SpringServerController.java
@@ -1,37 +1,0 @@
-/*
- * Copyright (C) 2024-2024 Sermant Authors. All rights reserved.
- *
- *   Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- */
-
-package io.demo.spring.server;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.time.LocalDateTime;
-
-/**
- * controller
- *
- * @author daizhenyu
- * @since 2024-07-02
- **/
-@RestController
-public class SpringServerController {
-    @RequestMapping("hello")
-    public String sayHello() {
-        LocalDateTime dateTime = LocalDateTime.now();
-        return "hello, the current time is " + dateTime;
-    }
-}
