--- xds-service-discovery-demo/xds-service-discovery/config/config.yaml
@@ -1,2 +1,0 @@
-xds.service.discovery:
-  upstreamServiceName: spring-server
