--- xds-service-discovery-demo/xds-service-discovery/xds-service-discovery-plugin/pom.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <groupId>io.sermant</groupId>
-        <artifactId>xds-service-discovery</artifactId>
-        <version>2.0.0</version>
-    </parent>
-
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>xds-service-discovery-plugin</artifactId>
-
-    <properties>
-        <output.basedir>${pom.basedir}/../../product/spring-client</output.basedir>
-        <package.plugin.name>xds-service-discovery</package.plugin.name>
-        <package.plugin.type>plugin</package.plugin.type>
-        <config.skip.flag>false</config.skip.flag>
-        <maven.compiler.source>8</maven.compiler.source>
-        <maven.compiler.target>8</maven.compiler.target>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>io.sermant</groupId>
-            <artifactId>sermant-agentcore-core</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>javax.servlet</groupId>
-            <artifactId>javax.servlet-api</artifactId>
-            <version>3.1.0</version>
-            <scope>provided</scope>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-shade-plugin</artifactId>
-            </plugin>
-        </plugins>
-    </build>
-</project>
