--- xds-service-discovery-demo/xds-service-discovery/xds-service-discovery-plugin/src/main/java/io/sermant/xds/service/discovery/TemplateConfig.java
@@ -1,41 +1,0 @@
-/*
- * Copyright (C) 2024-2024 Sermant Authors. All rights reserved.
- *
- *   Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- */
-
-package io.sermant.xds.service.discovery;
-
-import io.sermant.core.config.common.ConfigFieldKey;
-import io.sermant.core.config.common.ConfigTypeKey;
-import io.sermant.core.plugin.config.PluginConfig;
-
-/**
- * config
- *
- * @author daizhenyu
- * @since 2024-07-02
- **/
-@ConfigTypeKey("xds.service.discovery")
-public class TemplateConfig implements PluginConfig {
-    @ConfigFieldKey("upstreamServiceName")
-    private String upstreamServiceName;
-
-    public String getUpstreamServiceName() {
-        return upstreamServiceName;
-    }
-
-    public void setUpstreamServiceName(String upstreamServiceName) {
-        this.upstreamServiceName = upstreamServiceName;
-    }
-}
